Move app header components into own directory
